🗓️🔧：Reschedule A-Team Fortnightly Team Meetings
Today, @jorgebucaran and i settled on better timings
for A-Team meetings that work best for us; although this
could still be subject to debate, please feel free to discuss
it at our next meeting if preferences arise. We prefer this time!

That is 8:00 AM here in Miami & 9:00 PM over in Tokyo. 🙌

Refs: https://www.wolframalpha.com/input?i=8am+miami+time+to+japan+time

Signed-off-by: Derek Lewis <dereknongeneric@open.inf.is>