📄♻️：turn non-persons roster into html
HTML tables are not too much more work to maintain
than Markdown, but by having it as HTML, we are able
to have more control. Conversion of the other tables will
follow. Many HTML table closing tags are optional and
hard to maintain, so we have left them off for now.

Refs: https://html.spec.whatwg.org/multipage/tables.html#table-examples

Signed-off-by: Derek Lewis <dereknongeneric@open.inf.is>